Separate out security groups for NLB/public traffic
Previously we used one set for both the NLB listener and the gateway, which made permissions reasoning slightly challenging. This creates a new security group just for the LB listeners and the SDM gateway, which only allows traffic within itself.

This also lets us have the SDM gateway listen on a standard port (443) instead of 5000

Fixes Security Hub warnings about open ingress from CIDR 0.0.0.0/0 and an open listener on a non-standard port (5000)